+++ examples/counter-web/CHANGELOG.md
@@ -1,5 +1,14 @@
 # counter-web
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [6aa89e1]
+- Updated dependencies
+  - @caatinga/core@2.0.0
+  - @caatinga/client@2.0.0
+
 ## 0.1.2
 
 ### Patch Changes